Moved get_acquisition_times() to load_nifti.py
